[domain] move RemoteMetaData into module
