+++ domain/src/main/scala/net/kemitix/s3thorp/domain/RemoteMetaData.scala
@@ -1,6 +1,4 @@
-package net.kemitix.s3thorp
-
-import net.kemitix.s3thorp.domain.{LastModified, MD5Hash, RemoteKey}
+package net.kemitix.s3thorp.domain
 
 final case class RemoteMetaData(remoteKey: RemoteKey,
                                 hash: MD5Hash,
